+++ tests/Unit/CrudPanel/CrudPanelFieldsTest.php
@@ -281,11 +281,23 @@ class CrudPanelFieldsTest extends BaseCrudPanelTest
 
     public function testBeforeField()
     {
-        $this->markTestIncomplete('Not correctly implemented');
-
         $this->crudPanel->addFields($this->threeTextFieldsArray);
 
-        // TODO: fix the before field method not preserving field keys
+        $this->crudPanel->beforeField('field2');
+
+        $createKeys = array_keys($this->crudPanel->create_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[1]]);
+        $this->assertEquals(['field1', 'field3', 'field2'], $createKeys);
+
+        $updateKeys = array_keys($this->crudPanel->update_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[1]]);
+        $this->assertEquals(['field1', 'field3', 'field2'], $updateKeys);
+    }
+
+    public function testBeforeFieldFirstField()
+    {
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
+
         $this->crudPanel->beforeField('field1');
 
         $createKeys = array_keys($this->crudPanel->create_fields);
@@ -295,51 +307,58 @@ class CrudPanelFieldsTest extends BaseCrudPanelTest
         $updateKeys = array_keys($this->crudPanel->update_fields);
         $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[0]]);
         $this->assertEquals(['field3', 'field1', 'field2'], $updateKeys);
+    }
+
+    public function testBeforeFieldLastField()
+    {
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
+
+        $this->crudPanel->beforeField('field3');
+
+        $createKeys = array_keys($this->crudPanel->create_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $createKeys);
+
+        $updateKeys = array_keys($this->crudPanel->update_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $updateKeys);
     }
 
     public function testBeforeFieldCreateForm()
     {
-        $this->markTestIncomplete('Not correctly implemented');
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
 
-        $this->crudPanel->addFields($this->threeTextFieldsArray, 'create');
-
-        // TODO: fix the before field method not preserving field keys
-        $this->crudPanel->beforeField('field1');
+        $this->crudPanel->beforeField('field1', 'create');
 
         $createKeys = array_keys($this->crudPanel->create_fields);
         $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[0]]);
         $this->assertEquals(['field3', 'field1', 'field2'], $createKeys);
 
-        $this->assertEmpty($this->crudPanel->update_fields);
+        $updateKeys = array_keys($this->crudPanel->update_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $updateKeys);
     }
 
     public function testBeforeFieldUpdateForm()
     {
-        $this->markTestIncomplete('Not correctly implemented');
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
 
-        $this->crudPanel->addFields($this->threeTextFieldsArray, 'update');
+        $this->crudPanel->beforeField('field1', 'update');
 
-        // TODO: fix the before field method not preserving field keys
-        $this->crudPanel->beforeField('field1');
+        $createKeys = array_keys($this->crudPanel->create_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $createKeys);
 
         $updateKeys = array_keys($this->crudPanel->update_fields);
         $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[0]]);
         $this->assertEquals(['field3', 'field1', 'field2'], $updateKeys);
-
-        $this->assertEmpty($this->crudPanel->create_fields);
     }
 
     public function testBeforeFieldForDifferentFieldsInCreateAndUpdate()
     {
-        $this->markTestIncomplete('Not correctly implemented');
-
         $this->crudPanel->addFields($this->threeTextFieldsArray, 'create');
         $this->crudPanel->addFields($this->twoTextFieldsArray, 'update');
 
-        // TODO: fix the before field method not preserving field keys
-        // TODO: in the case of the create form, this will move the 'field3' field before the 'field1' field, but for
-        //       the update form, it will move the 'field2' field before the 'field1' field. fix by adding second
-        //       optional $form parameter with 'both' as a default value.
         $this->crudPanel->beforeField('field1');
 
         $createKeys = array_keys($this->crudPanel->create_fields);
@@ -365,11 +384,8 @@ class CrudPanelFieldsTest extends BaseCrudPanelTest
 
     public function testAfterField()
     {
-        $this->markTestIncomplete('Not correctly implemented');
-
         $this->crudPanel->addFields($this->threeTextFieldsArray);
 
-        // TODO: fix the after field method not preserving field keys
         $this->crudPanel->afterField('field1');
 
         $createKeys = array_keys($this->crudPanel->create_fields);
@@ -379,48 +395,58 @@ class CrudPanelFieldsTest extends BaseCrudPanelTest
         $updateKeys = array_keys($this->crudPanel->update_fields);
         $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[1]]);
         $this->assertEquals(['field1', 'field3', 'field2'], $updateKeys);
+    }
+
+    public function testAfterFieldLastField()
+    {
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
+
+        $this->crudPanel->afterField('field3');
+
+        $createKeys = array_keys($this->crudPanel->create_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $createKeys);
+
+        $updateKeys = array_keys($this->crudPanel->update_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $updateKeys);
     }
 
     public function testAfterFieldCreateForm()
     {
-        $this->markTestIncomplete('Not correctly implemented');
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
 
-        $this->crudPanel->addFields($this->threeTextFieldsArray, 'create');
-
-        // TODO: fix the after field method not preserving field keys
-        $this->crudPanel->afterField('field1');
+        $this->crudPanel->afterField('field1', 'create');
 
         $createKeys = array_keys($this->crudPanel->create_fields);
         $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[1]]);
         $this->assertEquals(['field1', 'field3', 'field2'], $createKeys);
 
-        $this->assertEmpty($this->crudPanel->update_fields);
+        $updateKeys = array_keys($this->crudPanel->update_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $updateKeys);
     }
 
     public function testAfterFieldUpdateForm()
     {
-        $this->markTestIncomplete('Not correctly implemented');
+        $this->crudPanel->addFields($this->threeTextFieldsArray);
 
-        $this->crudPanel->addFields($this->threeTextFieldsArray, 'update');
+        $this->crudPanel->afterField('field1', 'update');
 
-        // TODO: fix the after field method not preserving field keys
-        $this->crudPanel->afterField('field1');
+        $createKeys = array_keys($this->crudPanel->create_fields);
+        $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->create_fields[$createKeys[2]]);
+        $this->assertEquals(['field1', 'field2', 'field3'], $createKeys);
 
         $updateKeys = array_keys($this->crudPanel->update_fields);
         $this->assertEquals($this->expectedThreeTextFieldsArray['field3'], $this->crudPanel->update_fields[$updateKeys[1]]);
         $this->assertEquals(['field1', 'field3', 'field2'], $updateKeys);
-
-        $this->assertEmpty($this->crudPanel->create_fields);
     }
 
     public function testAfterFieldForDifferentFieldsInCreateAndUpdate()
     {
-        $this->markTestIncomplete('Not correctly implemented');
-
         $this->crudPanel->addFields($this->threeTextFieldsArray, 'create');
         $this->crudPanel->addFields($this->twoTextFieldsArray, 'update');
 
-        // TODO: fix the after field method not preserving field keys
         $this->crudPanel->afterField('field1');
 
         $createKeys = array_keys($this->crudPanel->create_fields);